requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
requests==2.31.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
pytest==7.4.0
pytest-cov==4.1.0
